Tweak encoder setup

+++ autonomous/system/vars.h
@@ -3,5 +3,5 @@ int last_power = 0;
 int left_encoder = 0;
 int prev_left_encoder = 0;
 
-int distance_counts = 0;
-int distance_feet = 0;
+long distance_counts = 0;
+float distance_feet = 0;
